+++ src/Audit.NET/AuditScope.cs
@@ -270,7 +270,7 @@ namespace Audit.Core
         /// Async version of the dispose method
         /// </summary>
         /// <returns></returns>
-#if NETSTANDARD2_1
+#if NETSTANDARD2_0 || NETSTANDARD2_1
         public async ValueTask DisposeAsync()
 #else
         public async Task DisposeAsync()